removed non-existent packets since they add a ton of clutter and may never see the day
