+++ Robot_stm32f4discovery/Src/robocup/hermes/hermes_task.c
@@ -56,6 +56,7 @@ void hermes_taskEntryPoint(void) {
 		}
 
 		// This is use to give back control to other task
+		// TODO: REMOVE!
 		osDelay(1);
 	}
 }
